perf: rework content tab
- file selection
- loading state
- keybind (enter to rename)
- auto refresh file content (update progress / priority)
- delay update when tab focuses instead of view
- persist opened items between updates / rename !!! NOT TESTED !!!
FIXME: auto refresh causes editing / newName to reset => rename canceled
